+++ app/templates/editor/edit.hbs
@@ -154,5 +154,6 @@
         closeNavMenu=(action "closeNavMenu")
         closeMenus=(action "closeMenus")
         deletePost=(action "toggleDeletePostModal")
+        updateSlug=updateSlug
     }}
 {{/liquid-wormhole}}
